Huge code refactoring
Huge code refactoring
-----
Make object framework for WinAPI and Graphics;
Tweaked physics a bit, slightly improved performance;
Tidied up Main.cpp;
Renamed components:
"TimeSystem" to "Tick",
"ObjectHandler" to "Objects";
Removed unused source files from project;
Code cleanup;

+++ Constants.h
@@ -1,5 +1,5 @@
 #pragma once
-typedef double CType;
+typedef float CType;
 const double EPS = 1e-24;
 const double G = 6.67448478 * 1e-11;
 const double SolarMass = 1.9885 * 1e30; //Kilogramm
